+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Search concern improvements
 - Date helper improvements
+- Add `boolean_icon` helper
 
 ## 1.0.15